cleanup arguments, allow us to specify a start mark
Often times we want to replay to the end of the mkfs, and then replay starting
from the mkfs while checking the fs every interval.  In order to do this change
the arguments so they corellate to what they actually do and add --start-mark so
we can skip to the mkfs mark and replay from that point onward with --fsck.

Signed-off-by: Josef Bacik <jbacik@fb.com>